Refactor batch and request helpers to util
